requirements.txt
dj-database-url==0.5.0
Django==2.2.2
django-authtools==1.6.0
django-dirtyfields==1.3.1
django-heroku==0.3.1
django-widget-tweaks==1.4.5
gunicorn==19.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
